Integration of Mock Generation with CI (#71)
* Move to BE folder

* Added to lahutil

* Added db seeding to CI

* Added coloring to text

* Separated out to new file
