+++ .github/workflows/deploy.yml
@@ -1,0 +1,17 @@
+name: Deployment Tasks
+
+on: [push]
+
+jobs:
+   
+  seed-db:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v1
+    - name: Setup CI Script 
+      run: bash scripts/setup_ci.sh
+      env: 
+        DB_URI: ${{ secrets.DB_URI }}
+    - name: Seed MongoDB Instance
+      run: docker-compose run backend node utils/generate_mock_data.js --color
+
