+++ assets/resource_template.json
@@ -1,0 +1,17 @@
+[
+  "{{repeat(100)}}",
+  {
+    "companyName": "{{company()}} {{company()}}",
+    "contactName": "{{firstName()}} {{surname()}}",
+    "contactPhone": "{{phone()}}",
+    "contactEmail": "{{email()}}",
+    "description": "{{lorem(3, \"sentences\")}}",
+    "address": "{{integer(100, 999)}} {{street()}}, {{city()}}, {{state()}}, {{integer(100, 10000)}}",
+    "location": {
+      "type": "Point",
+      "coordinates": ["{{floating(-180, 180)}}", "{{floating(-90, 90)}}"]
+    },
+    "notes": "{{lorem(2, \"sentences\")}}",
+    "tags": ["{{repeat(0,7)}}", "{{lorem(1, \"words\")}}"]
+  }
+]
