+++ backend/package.json
@@ -19,6 +19,7 @@
     "body-parser": "^1.19.0",
     "celebrate": "^10.0.1",
     "chai": "^4.2.0",
+    "colors": "^1.4.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
@@ -29,9 +30,11 @@
     "mocha": "^6.2.1",
     "mongoose": "^5.7.5",
     "morgan": "^1.9.1",
+    "node-fetch": "^2.6.0",
     "nodemon": "^1.19.3",
     "passport": "^0.4.0",
     "passport-google-oauth": "^2.0.0",
+    "progress": "^2.0.3",
     "sinon": "^7.5.0",
     "supertest": "^4.0.2"
   }
